requirements.txt
torch
transformers
fastapi[all]
uvicorn[standard]
sse-starlette
orjson
boto3
hnswlib
openai
openai-whisper @ git+https://github.com/openai/whisper
audiocraft @ git+https://github.com/facebookresearch/audiocraft@69fea8b290ad1b4b40d28f92d1dfc0ab01dbab85

TTS
pytube
pydub